requirements.txt
autogen-agentchat==0.4.5
autogen-ext[openai,azure]==0.4.5
beautifulsoup4==4.12.3
docxtpl==0.18.0
flaml[automl]==2.3.1
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
pydantic==2.10.6
llama-cloud==0.1.12
llama-cloud-services==0.6.0
llama-index==0.12.16
llama-index-agent-openai==0.4.5
llama-index-cli==0.4.0
llama-index-core==0.12.16.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.18
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-index-postprocessor-flag-embedding-reranker==0.3.0
llama-parse==0.6.0
streamlit==1.42.0
openpyxl==3.1.5
openai==1.61.1
# pandas==2.0.3
Pillow==11.1.0
PyMuPDF==1.25.3
python-docx==1.1.2
streamlit-option-menu==0.4.0
requests==2.31.0
selenium==4.26.1
lxml
